add standard completions
