+++ apps/jan-api-gateway/application/app/interfaces/http/routes/routes_provider.go
@@ -5,7 +5,8 @@ import (
 	v1 "menlo.ai/jan-api-gateway/app/interfaces/http/routes/v1"
 	"menlo.ai/jan-api-gateway/app/interfaces/http/routes/v1/auth"
 	"menlo.ai/jan-api-gateway/app/interfaces/http/routes/v1/auth/google"
-	"menlo.ai/jan-api-gateway/app/interfaces/http/routes/v1/chat"
+	chat "menlo.ai/jan-api-gateway/app/interfaces/http/routes/v1/chat"
+	conv_chat "menlo.ai/jan-api-gateway/app/interfaces/http/routes/v1/conv"
 	"menlo.ai/jan-api-gateway/app/interfaces/http/routes/v1/conversations"
 	"menlo.ai/jan-api-gateway/app/interfaces/http/routes/v1/mcp"
 	mcp_impl "menlo.ai/jan-api-gateway/app/interfaces/http/routes/v1/mcp/mcp_impl"
@@ -23,10 +24,12 @@ var RouteProvider = wire.NewSet(
 	organization.NewAdminApiKeyAPI,
 	organization.NewOrganizationRoute,
 	mcp_impl.NewSerperMCP,
-	chat.NewCompletionNonStreamHandler,
-	chat.NewCompletionStreamHandler,
-	chat.NewCompletionAPI,
 	chat.NewChatRoute,
+	chat.NewCompletionAPI,
+	conv_chat.NewConvChatRoute,
+	conv_chat.NewConvCompletionAPI,
+	conv_chat.NewCompletionNonStreamHandler,
+	conv_chat.NewCompletionStreamHandler,
 	mcp.NewMCPAPI,
 	v1.NewModelAPI,
 	responses.NewResponseRoute,
