+++ apps/jan-api-gateway/application/app/interfaces/http/routes/v1/conv/con_completion_nonstream_handler.go
@@ -1,4 +1,4 @@
-package chat
+package conv
 
 import (
 	"context"
